% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encSeq.R
\name{encSeq}
\alias{encSeq}
\title{Extract sequence from \code{GRanges} and generate encodings.}
\usage{
encSeq(GRanges, BSgenome, window = 0, type = "token")
}
\arguments{
\item{GRanges}{A \code{\link{GRanges}} object provides coordinates of interest.}

\item{BSgenome}{A \code{\link{BSgenome}} object containing the sequence of the genome.}

\item{window}{An integer value indicates the width of flanking region upstream and downstream.}

\item{type}{A character indicates the type of sequence encodings. Currently, only 'token' and 'onehot' are supported.}
}
\value{
A matrix contains the sequence encodings of regions from GRanges - window to GRanges + window.
}
\description{
A function used to get sequence features from \code{\link{GRanges}}
}
\examples{

## Generate sequence encoding for example data.
library(BSgenome.Hsapiens.UCSC.hg38)
input <- import.bed(system.file("extdata", "example.bed", package = "Geo2vec"))
encoding <- encSeq(input, BSgenome.Hsapiens.UCSC.hg38, window=250, type='token')

}
