% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo2df.R
\name{geo2df}
\alias{geo2df}
\title{Convert GRanges to data frame.}
\usage{
geo2df(GRanges, org_site_id = NULL)
}
\arguments{
\item{GRanges}{A \code{GRanges} object whose metadata columns are generated Geo2vec encodings.}

\item{org_site_id}{A set of index indicating which entries in the original input Geo2vec came from.}
}
\value{
A \code{data.frame} object contains the Geo2vec encodings.
}
\description{
This function converts \code{GRanges} objects containing Geo2vec encodings to data frames.
}
\examples{
\donttest{
## Convert chunkTX to dataframe for saving.

chunkTX <- site_chunkTX(input, txdb, exon_only = T, long_tx = T, mRNA = T)
out_df <- geo2df(chunkTX)
} 
}
