% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridTX.R
\name{gridTX}
\alias{gridTX}
\title{Generate gridTX for single-nucleotide level data.}
\usage{
gridTX(
  x,
  txdb,
  ngrid = 40,
  exon_only = TRUE,
  long_tx = TRUE,
  mRNA = TRUE,
  unify_strand = TRUE
)
}
\arguments{
\item{x}{A \code{GRanges} object}

\item{txdb}{A transcript database, currently only TxDb and EnsDb are supported.}

\item{ngrid}{An integer indicating the number of grids, i.e., how many pieces will the transcript be divided into.}

\item{exon_only}{A \code{logical} object indicating whether to consider only sites from exonic regions.}

\item{long_tx}{A \code{logical} object indicating whether to consider only the longest transcript for each site.}

\item{mRNA}{A \code{character} object indicating whether to consider only protein coding transcripts.}

\item{unify_strand}{A \code{logical} object indicating whether to unify the direction from 5'end to 3'end.}
}
\value{
A \code{\link{GRanges}} object whose metacolumns are the generated gridTX.
}
\description{
This function accepts GRanges data with width 1 and generate gridTX for each input.
}
\examples{

## Generate gridTX for example data.
library(EnsDb.Hsapiens.v86)
input <- import.bed(system.file("extdata", "example.bed", package = "Geo2vec"))
encoding <- gridTX(input+10, EnsDb.Hsapiens.v86, ngrid=40, exon_only=T, long_tx=T, mRNA=T)

}
