% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_chunkTX.R
\name{region_chunkTX}
\alias{region_chunkTX}
\title{Generate chunkTX for region level data.}
\usage{
region_chunkTX(x, txdb, exon_only = TRUE, long_tx = TRUE, mRNA = TRUE)
}
\arguments{
\item{x}{A \code{GRanges} object.}

\item{txdb}{A transcript database, currently only TxDb and EnsDb are supported.}

\item{exon_only}{A \code{logical} object indicating whether to consider only regions whose starts and ends are from exons.}

\item{long_tx}{A \code{logical} object indicating whether to consider only the longest transcript for each region.}

\item{mRNA}{A \code{character} object indicating whether to consider only protein coding transcripts.}
}
\value{
A \code{\link{GRanges}} object whose metacolumns are the generated chunkTX.
}
\description{
This function accepts \code{GRanges} data with width larger than 1 and generate chunkTX for each input.
}
\examples{

## Generate chunkTX for example data.
library(EnsDb.Hsapiens.v86)
input <- import.bed(system.file("extdata", "example.bed", package = "Geo2vec"))
encoding <- region_chunkTX(input+10, EnsDb.Hsapiens.v86, exon_only=T, long_tx=T, mRNA=T)

}
