% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSequence.R
\name{sampleSequence}
\alias{sampleSequence}
\title{Sample motifs on a subset of genome.}
\usage{
sampleSequence(
  motif,
  region,
  bsgenome,
  N = NULL,
  fixed = FALSE,
  replace = FALSE
)
}
\arguments{
\item{motif}{A character string indicating the query sequence or motifs; Vague mapping rules in \code{\link{IUPAC_CODE_MAP}} is supported when \code{fixed} = FALSE.}

\item{region}{The \code{\link{GRanges}} object defines the subset region of the genome.}

\item{N}{Number of ranges sampled, by default, it returns all the matched ranges without sub-sampling.}

\item{fixed}{FALSE to support the vague mapping rule of the character string, default is FALSE.}

\item{replace}{Whether sample with replacement or not, default is FALSE.}

\item{sequence}{The \code{\link{BSgenome}} object containing the sequence of the genome.}
}
\value{
A \code{GRanges} object contains the (sampled) mapped regions of the query sequence on the given subset of the genome.
}
\description{
A function used to extract and sample given motifs from a user defined sub-regions of the genome (Inherited from github.com/ZW-xjtlu/RegionPropertiesFeatures)
}
\examples{

## Sample DRACH motif on a given range:
library(BSgenome.Hsapiens.UCSC.hg38)
exons_gr <- GRanges(seqnames = c('chr10', 'chr19'),
                    IRanges(start = c(73813868, 3976270),
                            end = c(73813908, 3976310)),
                    strand = c('-', '-'))
motif_gr <- sampleSequence('DRACH', exons_gr, BSgenome.Hsapiens.UCSC.hg38, fixed = F)


}
