% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_chunkTX.R
\name{site_chunkTX}
\alias{site_chunkTX}
\title{Generate chunkTX for single-nucleotide level data.}
\usage{
site_chunkTX(x, txdb, exon_only = TRUE, long_tx = TRUE, mRNA = TRUE)
}
\arguments{
\item{x}{A \code{GRanges} object}

\item{txdb}{A transcript database, currently only TxDb and EnsDb are supported.}

\item{exon_only}{A \code{logical} object indicating whether to consider only sites from exonic regions.}

\item{long_tx}{A \code{logical} object indicating whether to consider only the longest transcript for each site.}

\item{mRNA}{A \code{character} object indicating whether to consider only protein coding transcripts.}
}
\value{
A \code{\link{GRanges}} object whose metacolumns are the generated chunkTX.
}
\description{
This function accepts GRanges data with width 1 and generate chunkTX for each input.
}
\examples{

## Generate chunkTX for example data.
library(EnsDb.Hsapiens.v86)
input <- import.bed(system.file("extdata", "example.bed", package = "Geo2vec"))
encoding <- site_chunkTX(input, EnsDb.Hsapiens.v86, exon_only=T, long_tx=T, mRNA=T)

}
