% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_landmarkTX.R
\name{site_landmarkTX}
\alias{site_landmarkTX}
\title{Generate landmarkTX for single-nucleotide level data.}
\usage{
site_landmarkTX(x, txdb, long_tx = TRUE, mRNA = TRUE)
}
\arguments{
\item{x}{A \code{GRanges} object}

\item{txdb}{A transcript database, currently only TxDb and EnsDb are supported.}

\item{long_tx}{A \code{logical} object indicating whether to consider only the longest transcript for each site.}

\item{mRNA}{A \code{character} object indicating whether to consider only protein coding transcripts.}
}
\value{
A \code{\link{GRanges}} object whose metacolumns are the generated landmarkTX.
}
\description{
This function accepts GRanges data with width 1 and generate landmarkTX for each input.
}
\examples{

## Generate landmarkTX for example data.
library(EnsDb.Hsapiens.v86)
input <- import.bed(system.file("extdata", "example.bed", package = "Geo2vec"))
encoding <- site_landmarkTX(input, EnsDb.Hsapiens.v86, long_tx=T, mRNA=T)

}
