% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_onehotTX.R
\name{site_onehotTX}
\alias{site_onehotTX}
\title{Generate onehotTX for single-nucleotide level data.}
\usage{
site_onehotTX(
  x,
  txdb,
  window = 250,
  exon_only = TRUE,
  long_tx = TRUE,
  mRNA = TRUE
)
}
\arguments{
\item{x}{A \code{GRanges} object}

\item{txdb}{A transcript database, currently only TxDb and EnsDb are supported.}

\item{window}{An integer indicates the width of flanking region which equals to 2*window + 1.}

\item{exon_only}{A \code{logical} object indicating whether to consider only sites from exonic regions.}

\item{long_tx}{A \code{logical} object indicating whether to consider only the longest transcript for each site.}

\item{mRNA}{A \code{character} object indicating whether to consider only protein coding transcripts.}
}
\value{
A \code{\link{GRanges}} object whose metacolumns are the generated onehotTX.
}
\description{
This function accepts GRanges data with width 1 and generate onehotTX for each input.
}
\examples{

## Generate onehotTX for example data.
library(EnsDb.Hsapiens.v86)
input <- import.bed(system.file("extdata", "example.bed", package = "Geo2vec"))
encoding <- site_onehotTX(input, EnsDb.Hsapiens.v86, window=250, exon_only=T, long_tx=T, mRNA=T)

}
